Use file scoped namespace
